update alacritty config with new defaults
Alacritty used to generate a default config file, but now it just ships
with an example file as documentation.  That file has most things
commented out, so I copied most of that hoping that the program defaults
would remain the same if they are commented out in the config.  If there
are behavior changes I'll have to find and uncomment some of these.