rearrange vim files
This primarily moves vim's config files to ./config/vim and plugin
storage to the cache. Vim has supported reading config files from
`XDG_CONFIG_HOME` for a while now, so this finally moves them.

The former ~/.vim directory contains only state that can be
automatically regenerated, so that is being moved to a cache directory
to keep it out of the home directory and backups. This requires updating
vim's runtime path. The viminfo file is also moved to the cache
directory because losing it is not a big deal.

The vimrc file used to have a check before sourcing the vim-plug config
file, but they now live next to each other in the repo and are deployed
together, so that check seems unnecessary and was removed.